NavigationBarSettings: User toggle for navbar visibility.
** CleanUp Buttons Fragment filled with NavBar Shit
