+++ res/xml/x_settings_button.xml
@@ -32,13 +32,6 @@
             android:title="@string/hardware_keys_disable_title"
             android:summary="@string/hardware_keys_disable_summary" />
 
-        <SwitchPreference
-            android:key="disable_nav_keys"
-            android:title="@string/disable_navkeys_title"
-            android:summary="@string/disable_navkeys_summary"
-            android:persistent="false"
-            android:defaultValue="false" />
-
         <com.msm.xtended.preferences.SystemSettingSwitchPreference
             android:key="swap_navigation_keys"
             android:title="@string/swap_navigation_keys_title"
