+++ res/xml/x_settings_navigation.xml
@@ -18,6 +18,13 @@
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:settings="http://schemas.android.com/apk/res/com.android.settings" >
 
+    <SwitchPreference
+        android:key="enable_nav_bar"
+        android:title="@string/enable_navbar_title"
+        android:summary="@string/enable_navbar_summary"
+        android:defaultValue="true"
+        android:persistent="false" />
+
     <Preference
             android:title="@string/navbar_tuner_title" >
             <intent android:action="android.intent.action.MAIN"
