+++ src/com/msm/xtended/fragments/ButtonSettings.java
@@ -55,7 +55,6 @@ public class ButtonSettings extends ActionFragment implements OnPreferenceChange
 
     private static final String HWKEY_DISABLE = "hardware_keys_disable";
     private static final String KEY_BUTTON_SWAP_KEYS = "swap_navigation_keys";
-    private static final String DISABLE_NAV_KEYS = "disable_nav_keys";
 
     // category keys
     private static final String CATEGORY_HWKEY = "hardware_keys";
@@ -82,9 +81,6 @@ public class ButtonSettings extends ActionFragment implements OnPreferenceChange
     private CustomSeekBarPreference mButtonBrightness;
     private SwitchPreference mButtonBrightness_sw;
     private SystemSettingSwitchPreference mSwapKeysPreference;
-    private SwitchPreference mDisableNavigationKeys;
-    private boolean mIsNavSwitchingMode = false;
-    private Handler mHandler;
 
     @Override
     public void onCreate(Bundle icicle) {
@@ -94,20 +90,6 @@ public class ButtonSettings extends ActionFragment implements OnPreferenceChange
         final Resources res = getResources();
         final ContentResolver resolver = getActivity().getContentResolver();
         final PreferenceScreen prefScreen = getPreferenceScreen();
-
-        // Force Navigation bar related options
-        mDisableNavigationKeys = (SwitchPreference) findPreference(DISABLE_NAV_KEYS);
-
-        // Only visible on devices that does not have a navigation bar already
-        if (ActionUtils.isHWKeysSupported(getActivity())) {
-            mDisableNavigationKeys.setOnPreferenceChangeListener(this);
-            mHandler = new Handler();
-            // Remove keys that can be provided by the navbar
-            updateDisableNavkeysOption();
-            updateDisableNavkeysCategories(mDisableNavigationKeys.isChecked());
-        } else {
-            prefScreen.removePreference(mDisableNavigationKeys);
-        }
 
         final boolean needsNavbar = ActionUtils.hasNavbarByDefault(getActivity());
         final PreferenceCategory hwkeyCat = (PreferenceCategory) prefScreen
@@ -272,69 +254,8 @@ public class ButtonSettings extends ActionFragment implements OnPreferenceChange
                     value ? 1 : 0);
             setActionPreferencesEnabled(!value);
             return true;
-        } else if (preference == mDisableNavigationKeys) {
-            if (mIsNavSwitchingMode) {
-                return false;
-            }
-            mIsNavSwitchingMode = true;
-            boolean isChecked = ((Boolean) newValue);
-            mDisableNavigationKeys.setEnabled(false);
-            writeDisableNavkeysOption(isChecked);
-            updateDisableNavkeysOption();
-            updateDisableNavkeysCategories(true);
-            mHandler.postDelayed(new Runnable() {
-                @Override
-                public void run() {
-                    mIsNavSwitchingMode = false;
-                }
-            }, 1500);
-            return true;
         }
         return false;
-    }
-
-    private void writeDisableNavkeysOption(boolean enabled) {
-        Settings.System.putIntForUser(getActivity().getContentResolver(),
-                Settings.System.FORCE_SHOW_NAVBAR, enabled ? 1 : 0, UserHandle.USER_CURRENT);
-    }
-
-    private void updateDisableNavkeysOption() {
-        boolean enabled = Settings.System.getIntForUser(getActivity().getContentResolver(),
-                Settings.System.FORCE_SHOW_NAVBAR, 0, UserHandle.USER_CURRENT) != 0;
-
-        mDisableNavigationKeys.setChecked(enabled);
-    }
-
-    private void updateDisableNavkeysCategories(boolean navbarEnabled) {
-        final PreferenceScreen prefScreen = getPreferenceScreen();
-
-        /* Disable hw-key options if they're disabled */
-        final PreferenceCategory homeCategory =
-                (PreferenceCategory) prefScreen.findPreference(CATEGORY_HOME);
-        final PreferenceCategory backCategory =
-                (PreferenceCategory) prefScreen.findPreference(CATEGORY_BACK);
-        final PreferenceCategory menuCategory =
-                (PreferenceCategory) prefScreen.findPreference(CATEGORY_MENU);
-        final PreferenceCategory assistCategory =
-                (PreferenceCategory) prefScreen.findPreference(CATEGORY_ASSIST);
-        final PreferenceCategory appSwitchCategory =
-                (PreferenceCategory) prefScreen.findPreference(CATEGORY_APPSWITCH);
-
-        if (homeCategory != null) {
-            homeCategory.setEnabled(!navbarEnabled);
-        }
-        if (backCategory != null) {
-            backCategory.setEnabled(!navbarEnabled);
-        }
-        if (menuCategory != null) {
-            menuCategory.setEnabled(!navbarEnabled);
-        }
-        if (assistCategory != null) {
-            assistCategory.setEnabled(!navbarEnabled);
-        }
-        if (appSwitchCategory != null) {
-            appSwitchCategory.setEnabled(!navbarEnabled);
-        }
     }
 
     @Override
